requirements.txt
black==23.9.1
flake8==6.1.0
flake8-quotes==3.3.2
flake8-variables-names==0.0.6
isort==5.12.0
pep8-naming==0.13.3
requests==2.31.0
python-slugify==8.0.1
